MAINT Rename SpectralEmbedding --> LaplacianEigenmap
