+++ .mailmap
@@ -46,6 +46,7 @@ Gilles Louppe <g.louppe@gmail.com> <g.louppe@ulg.ac.be>
 Hamzeh Alsalhi <93hamsal@gmail.com>
 Harikrishnan S <hihari777@gmail.com>
 Hendrik Heuer <hendrikheuer@gmail.com>
+Henry Lin <hlin117@gmail.com>
 Hrishikesh Huilgolkar <hrishikesh911@gmail.com> <hrishikesh@QE-IND-WKS007.(none)>
 Immanuel Bayer <mane.desk@gmail.com>
 Jacob Schreiber <jmschreiber91@gmail.com>
