+++ sklearn/manifold/__init__.py
@@ -7,6 +7,7 @@ from .isomap import Isomap
 from .mds import MDS
 from .spectral_embedding_ import SpectralEmbedding, spectral_embedding
 from .t_sne import TSNE
+from .laplacian_eigenmap_ import LaplacianEigenmap, laplacian_eigenmap
 
 __all__ = ['locally_linear_embedding', 'LocallyLinearEmbedding', 'Isomap',
-           'MDS', 'SpectralEmbedding', 'spectral_embedding', "TSNE"]
+           'MDS', "TSNE", "LaplacianEigenmap", "laplacian_eigenmap"]
